### 2024-02-19 #6381 @editorialbot
[PRE REVIEW]: gollum: An intuitive programmatic and visual interface for precomputed synthetic spectral model grids
**Submitting author:** @Sujay-Shankar (<a href="http://orcid.org/0000-0002-2290-6810">Sujay Shankar</a>)
**Repository:** https://github.com/BrownDwarf/gollum
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.0
