### 2024-04-10 #6381 @warrickball
Is the author's GitHub username @Sujay-Shankar or @SujayShankar2001? I suspect the former but the metadata has been populated with the latter. @Sujay-Shankar, could you let us know which is correct? I think this is generated from your JOSS profile, so you might want to check that you've put the correct username there.
