### 2024-04-10 #6381 @Sujay-Shankar
Hi @warrickball , thanks for picking up our paper for review! Sorry about the name mismatch, I recently changed my GitHub username. Sujay-Shankar is the new and current one, SujayShankar2001 is the old one. I just went into my JOSS profile and updated the username, thanks for reminding me!
