### 2024-04-25 BrownDwarf/gollum#103 @rhdtownsend
Issues with "python setup.py develop"
During the final step of the (developer) installation instructions (https://gollum-astro.readthedocs.io/en/latest/install.html), I run into some deprecation warnings and then an error:

```
running develop
/Users/townsend/miniconda3/envs/gollum_dev/lib/python3.8/site-packages/setuptools/command/develop.py:40: EasyInstallDeprecationWarning: easy_install command is deprecated.
!!

        ********************************************************************************
        Please avoid running ``setup.py`` and ``easy_install``.
        Instead, use pypa/build, pypa/installer or other
        standards-based tools.

        See https://github.com/pypa/setuptools/issues/917 for details.
        ********************************************************************************

!!
  easy_install.initialize_options(self)
/Users/townsend/miniconda3/envs/gollum_dev/lib/python3.8/site-packages/setuptools/_distutils/cmd.py:66: SetuptoolsDeprecationWarning: setup.py install is deprecated.
!!

        ********************************************************************************
        Please avoid running ``setup.py`` directly.
        Instead, use pypa/build, pypa/installer or other
        standards-based tools.

        See https://blog.ganssle.io/articles/2021/10/setup-py-deprecated.html for details.
        ********************************************************************************

!!
  self.initialize_options()
running egg_info
creating src/gollum.egg-info
writing src/gollum.egg-info/PKG-INFO
writing dependency_links to src/gollum.egg-info/dependency_links.txt
writing requirements to src/gollum.egg-info/requires.txt
writing top-level names to src/gollum.egg-info/top_level.txt
writing manifest file 'src/gollum.egg-info/SOURCES.txt'
reading manifest file 'src/gollum.egg-info/SOURCES.txt'
adding license file 'LICENSE'
writing manifest file 'src/gollum.egg-info/SOURCES.txt'
running build_ext
Creating /Users/townsend/miniconda3/envs/gollum_dev/lib/python3.8/site-packages/gollum.egg-link (link to src)
Adding gollum 0.3.0 to easy-install.pth file

Installed /Users/townsend/devel/gollum/src
Processing dependencies for gollum==0.3.0
error: astropy 5.1 is installed but astropy!=5.1.0,>=5.0.6 is required by {'ndcube'}
```

Suggest migration away from setup.py and easy_install. For the dependency error, why isn't a compatible astropy installed into the environment?

Review link: https://github.com/openjournals/joss-reviews/issues/6601
