### 2024-04-25 BrownDwarf/gollum#103 @Sujay-Shankar
Hi @rhdtownsend ! Thanks for your feedback on the setup process! The environment.yml file seems to have flown under the radar; it seems to be a little outdated, as is the setup.py, I'll look into updating the setup to use something newer as soon as possible.
The astropy errors are a bit cryptic, I'd assume astropy 5.1 is the same as 5.1.0, but gollum doesn't enforce astropy to be 5.1.0 anywhere as far as I know. And any 5.1 version should be greater than 5.0.6 so ndcube shouldn't be throwing any errors either. This could be some weirdness with Python 3.8, or hopefully just the setup process in general causing these odd errors. Either way, I'll test these things out and get back to you as soon as I can with the update/fix. Thanks again!
