### 2024-04-26 05c1ec2
Update setup process and documentation
