### 2024-04-30 e5fa7c5
Eliminate extraneous warnings in test_precomputed
Closes #104
