### 2024-04-30 BrownDwarf/gollum#103 @Sujay-Shankar
Small update on this: The documentation has been updated to switch to using pip instead of setup.py to create a development environment for gollum. I'm waiting on some things with @gully and readthedocs for the changes to be rendered on the online documentation, but the errors you've brought up have successfully been fixed! I'll close this issue once the new documentation goes live.
