+++ .readthedocs.yml
@@ -1,0 +1,13 @@
+version: 0.3.0
+
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.12"
+
+sphinx:
+  configuration: docs/conf.py
+
+python:
+  install:
+    - requirements: requirements.txt
