### 2024-05-02 ac70997
Fix config file version (2, not 0.3.0)
