+++ .readthedocs.yml
@@ -1,4 +1,4 @@
-version: 0.3.0
+version: 2
 
 build:
   os: ubuntu-22.04
