+++ requirements.txt
@@ -1,4 +1,3 @@
-python>=3.8
 numpy
 scipy
 matplotlib
