### 2024-05-02 82cc44d
Add sphinx-gallery to requirements
