### 2024-05-02 641b615
Add sphinx-material to requirements
