### 2024-05-02 BrownDwarf/gollum#103 @Sujay-Shankar
@rhdtownsend The documentation is now live on readthedocs and the revamped setup process is now available! I'll close the issue as completed now. Thanks again for bringing this to our attention!
