### 2024-05-10 BrownDwarf/gollum#105 @DanielAndreasen
Configuration for running the code
Hi all,

It's a very nice tool you have put together here, and I wish it was available when I did my PhD.

I think the paths for the libraries we have to download should be in a configuration file, rather than hardcoded into the software itself (it can default to what you have now though: `~/libraries`).
But if you have a file like `~/.config/gollum/config.yml` (or whichever format you prefer), where the path to the libraries can be set, and possible other settings as well, I think it would be very nice.

Review link: https://github.com/openjournals/joss-reviews/issues/6601

Sorry for the late review, but more will come soon as I'm getting into testing.
