### 2024-05-10 BrownDwarf/gollum#105 @Sujay-Shankar
Hi Daniel:

Thanks for your feedback! You are absolutely right that a configuration file is the way to go, and while the main gollum library doesn't have it yet, we're in the process of writing up a V2 interface that uses I believe .env files to house path data for grids.
I'll get working on implementing a similar thing for the original interface, and have it point to the same directory.
