+++ src/gollum/precomputed_spectrum.py
@@ -11,7 +11,9 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from copy import deepcopy
+from dotenv import get_key
 from logging import getLogger
+from pathlib import Path
 from warnings import filterwarnings
 from gollum.utilities import apply_numpy_mask
 from scipy.ndimage import gaussian_filter1d
