### 2024-05-10 BrownDwarf/gollum#105 @Sujay-Shankar
Updates: With the last two commits, this issue should be solved. I've made PHOENIX, Sonora, and coolTLUSTY classes now default to pulling their path from config.env, located in gollum's source directory. The actual values are the same as before. The setup files have also been updated to reflect the new dependency of python-dotenv. Please let me know if there's anything else that comes up pertaining to this issue, if not we should be able to safely close this one.
