### 2024-05-11 BrownDwarf/gollum#105 @DanielAndreasen
Great work. The only thing a bit annoying might be that changing the config, means the repository is out of sync, which can lead to problems when doing `git pull`, but I can live with that. Alternatively you could just write in the `README.md` how to write the config file, and have it in `.gitignore`, but this is also fine. Up to you if you will change it, or just close the issue.
