### 2024-05-11 BrownDwarf/gollum#106 @DanielAndreasen
Quickstart tutoirial - without jupyter notebook
Hi again,

I've tried running the quickstart tutorial (https://gollum-astro.readthedocs.io/en/latest/quickstart.html) from ipython (not notebook), and I don't get a plot after running the commands:
```python
ax = normalized_spectrum.plot()
ax.set_ylim(0, 2)
```

It might be my setup, but are you able to do so? Do I need to change something compared to the jupyter notebook code?

I am able to produce the plots without problems in the notebook though.

Review link: https://github.com/openjournals/joss-reviews/issues/6601
