### 2024-05-11 BrownDwarf/gollum#106 @DanielAndreasen
I solved it myself. I had to install an extra dependency: `pip install PyQt5`, and then run the code like:
```python
import matplotlib.pyplot as plt

from gollum.phoenix import PHOENIXSpectrum

spec = PHOENIXSpectrum(teff=5000, logg=4)
normalized_spectrum = spec.normalize()
ax = spec.normalize().plot()
ax.set_ylim(0)
plt.show()
```

This might just be my system, but maybe this issue can help others.
