### 2024-05-11 BrownDwarf/gollum#107 @DanielAndreasen
Formatting of math in tutorial
Some weird formatting in the first paragraph here, where you want to write `vsin(i)`:
https://gollum-astro.readthedocs.io/en/latest/tutorials/best_fit_for_fixed_template.html

Review link: https://github.com/openjournals/joss-reviews/issues/6601
