### 2024-05-11 BrownDwarf/gollum#108 @DanielAndreasen
Rotational fitting
This tutorial could use some polishing:
https://gollum-astro.readthedocs.io/en/latest/tutorials/best_fit_for_fixed_template.html

0. Some weird formatting in the first paragraph here, where you want to write vsin(i) (closing #107)
1. It feels like it assumes the user already have `muler` installed. Just add a small paragraph saying it can be installed with `pip install muler`, so the user don't have to find this information.
2. In code block 3, where you define the path to the spectrum, I suggest using a more descriptive path. Maybe something like `~/Downloads/Goldilocks_20210517T054403_v1.0_0060.spectra.fits`.
3. Not at all necessary, but the `data_clean` function can be writte quite elegantly (in my opinion) like below (I like to use type annotations), since all the methods used on the data, return the same type. But I do recognize that for some users this might be confusing.
4. Code block 8 and other places. Semi-colons are not illegal at the end of a line in python, but they should be.
5. Code block 9: I believe `metallicity` should be replaced with `Z`.
6. Code block 15: the `chi2` calculations gives me a bunch of `inf`. Should it be: `chi2_loss = np.sum(0.5 * residual.flux**2) / np.sum(residual.uncertainty.array**2)`? If I change it to that, the code works for me and give the same results as written in the tutorial.

```python
def data_clean(data: HPFSpectrumList) -> HPFSpectrum:
    """Clean the HPF data with standard post-processing techniques"""
    return (
        data.sky_subtract(method="vector")
        .deblaze()
        .trim_edges((4, 2942))
        .normalize()
        .stitch()
    )
```

Review link: https://github.com/openjournals/joss-reviews/issues/6601
