### 2024-05-11 BrownDwarf/gollum#107 @DanielAndreasen
Transferred to other issue, together with other minor details.
