### 2024-05-14 #6601 @warrickball
Hi @rhdtownsend, it looks like @Sujay-Shankar has fixed both the issues you've raised. Were there any new issues you've run into?

By the way, this approach of raising code issues in the repo is great. Thanks for consistently tagging this review thread (@DanielAndreasen too!).
