### 2024-05-14 BrownDwarf/gollum#105 @Sujay-Shankar
Thanks for pointing that out, you're definitely right. I'll go with a 'constant' template config file, and create some mechanism to automatically copy the template to a git untracked config.env if it doesn't exist, and not overwrite existing config otherwise. I'll close the issue once I push this change.
