### 2024-05-14 BrownDwarf/gollum#106 @Sujay-Shankar
I agree with you, our documentation seems to be missing the actual show method for the chart. I think this relies on notebooks' behavior of rendering certain objects when expressed as a literal, but IPython not necessarily doing the same. I'll make this fix and close when complete.
