### 2024-05-14 BrownDwarf/gollum#108 @Sujay-Shankar
Thanks for the feedback on the docs! Admittedly they're quite old and have flown under my radar until now, but all your points make sense. I'll address as many as possible and let you know how it goes.
