### 2024-05-14 BrownDwarf/gollum#110 @rhdtownsend
Change directory where gollum searches for grids
Is there a way to change where gollum searches for grid data? If it is hardwired to ~/libraries/raw, that should be user-changeable.

Review link: https://github.com/openjournals/joss-reviews/issues/6601
