### 2024-05-14 BrownDwarf/gollum#111 @rhdtownsend
Docs make presumption about download directory
A number of places in the docs presume that the download directory will be ~/Downloads (see, e.g., the instructions [here](https://gollum-astro.readthedocs.io/en/latest/fetching_models.html). While that may be true on MacOS, it certainly isn't part of the [Linux FHS](https://en.wikipedia.org/wiki/Filesystem_Hierarchy_Standard). Suggest making the docs more download-agnostic. For instance, in the case of the Phoenix download instructions, why not have the user directly download into their chosen spectral library directory, rather than to ~/Downloads?

Review link: https://github.com/openjournals/joss-reviews/issues/6601
