### 2024-05-14 BrownDwarf/gollum#112 @rhdtownsend
Comparison against MSG
It would be helpful for the documentation to compare gollum's capabilities against those of [MSG](https://github.com/rhdtownsend/msg) (full disclosure, I'm the developer of MSG). From what I can see, gollum does not support any kind of interpolation in grids (MSG does). Is this a design choice, made under the expectation that higher-level modules will provide interpolation functionality? If so, it would be helpful having that narrative in the introductory documentation, so users can quickly assess whether gollum is the right fit (pardon the pun!) for them.

Review link: https://github.com/openjournals/joss-reviews/issues/6601
