### 2024-05-14 BrownDwarf/gollum#113 @rhdtownsend
API documents need cleaning up
The API docs have seen quite a bit of improvement since I began reviewing gollum, but still need some tidying up. For instance, there is a bogus "PHOENIX Spectrum" header at the top of the class list.

Review link: https://github.com/openjournals/joss-reviews/issues/6601
