### 2024-05-14 BrownDwarf/gollum#110 @DanielAndreasen
Hi @rhdtownsend, I think this is related to #105.
