### 2024-05-14 c8a6306
Polish up config
- Add template config
- Untrack true config
- Add mechanism where if true config is not found, to copy template config
