### 2024-05-14 BrownDwarf/gollum#110 @Sujay-Shankar
Hi all, I believe the most recent commit has finished up the path issue, gollum now automatically creates your config with default paths if it doesn't detect one (located in gollum's source directory, all you have to do to kick off this auto-creation is import PrecomputedSpectrum, or any module that subclasses it such as PHOENIX or Sonora), and this config file can be changed to a new user default without further git stuff overwriting it. If there is anything here that wasn't mentioned in #105 , do let me know, otherwise we should be good to close this one.
