### 2024-05-14 BrownDwarf/gollum#110 @Sujay-Shankar
I also will be adding instructions regarding the config system to the documentation when I get around to editing that.
