### 2024-05-14 BrownDwarf/gollum#106 @Sujay-Shankar
Okay, I believe I've updated our quickstart tutorial. Regarding pyqt5, I'm trying to reproduce the error, but the code is running fine, both in notebook and in raw IPython, without a need for pyqt5 to be explicitly installed. Did you use one of the two recommended environment creation methods in the updated documentation?
