### 2024-05-15 BrownDwarf/gollum#106 @DanielAndreasen
I'm no longer a fan of conda, so I installed it using pip in a virtual environment using the `requirements.txt` file. Technically I used [`uv`](https://github.com/astral-sh/uv) which seems to become the standard, but that is just `pip` on steroids. It's fine if you close this issue, but it can be used for others if they have similar problems.
