### 2024-05-15 BrownDwarf/gollum#106 @Sujay-Shankar
I see, I myself don't technically use conda either, but mamba, which is to conda as uv is to pip, I assume. Unfortunately, pyqt5 is the pip package name while pyqt is the conda package name, meaning if I put it in requirements.txt it would throw an error for one of the two platforms. When you got the error, did it itself tell you to install pyqt5? If so, I think we can just leave it be for the time being, users will most likely just follow the error's directions. If it didn't explicitly tell you to install it I can add a note to the docs saying pip/uv users need to also manually install pyqt5 so they don't have to comb through stackoverflow for the fix.
