### 2024-05-15 BrownDwarf/gollum#108 @Sujay-Shankar
Ok, I think I've solved all the issues here:
1. Math formatting inside italics doesn't work, so I got rid of the italics.
2. I downloaded the data into gollum to both shorten the path and take the burden off the user for wrangling with path stuff
3. I removed the data clean function, it was a one-off, there wasn't really a need for it to be a function in the first place.
4. I removed all the semi-colons and put in our favorite plt.show()
5. metallicity was replaced with Z
6. The formula we used previously as far as I know doesn't look like the chi-square formula, and the description was sum of squares of residuals, which also isn't chi-square, so I switched the whole thing to using RSS (results are the same, and it's much simpler)

Aside from those, just generally spruced up the page with modern stuff like f-strings, concise numpy stuff, and cleaner figures.
