+++ src/gollum/precomputed_spectrum.py
@@ -15,7 +15,7 @@ from copy import deepcopy
 from dotenv import get_key
 from logging import getLogger
 from pathlib import Path
-from warnings import filterwarnings
+from warnings import filterwarnings, catch_warnings
 from gollum.utilities import apply_numpy_mask
 from scipy.ndimage import gaussian_filter1d
 from scipy.signal import find_peaks
@@ -115,12 +115,13 @@ class PrecomputedSpectrum(Spectrum1D):
         """
         lam0 = np.median(self.wavelength.value)
         x2 = (299792.458 * (self.wavelength.value - lam0) / (lam0 * vsini)) ** 2
-        kernel = np.where(
-            x2 < 1,
-            np.pi / 2 * u1 * (1 - x2)
-            + np.sqrt(1 - x2) * (2 - 2 * u1 - 4 / 3 * u2 * u2 * x2),
-            0,
-        )
+        with catch_warnings(action="ignore", category=RuntimeWarning):
+            kernel = np.where(
+                x2 < 1,
+                np.pi / 2 * u1 * (1 - x2)
+                + np.sqrt(1 - x2) * (2 - 2 * u1 - 4 / 3 * u2 * u2 * x2),
+                0,
+            )
         kernel, positive_elements = kernel / np.sum(kernel, axis=0), kernel > 0
         return (
             self._copy(
