### 2024-05-15 BrownDwarf/gollum#109 @Sujay-Shankar
I've dealt with the warnings in the recent commit, and yes, we download the spectra from the PHOENIX server. While we currently don't support or claim to support downloading and explicitly storing PHOENIX spectra, there is some caching behavior that astropy does when we use their IO API, so calling download on the same spectrum again is faster due to their hidden machinery.
