+++ docs/fetching_models.rst
@@ -86,3 +86,9 @@ Finally, you must copy the wavelength file into this directory as well.  Notice 
     ls ~/libraries/raw/PHOENIX/
     WAVE_PHOENIX-ACES-AGSS-COND-2011.fits  Z-0.0  Z+0.5  Z-0.5  Z+1.0  Z-1.0  Z-1.5  Z-2.0  Z-3.0  Z-4.0
 
+You can optionally set user defaults for gollum for each model grid. First, go into `gollum`'s source directory and look for a file called `config.env` (NOT `config_template.env`, do not edit this file!). If it doesn't exist, then run the following:
+
+    cp config_template.env config.env
+
+Now with `config.env` generated, you can edit it. Change the key corresponding to the model grid whose path you want to edit. Then save the file, and from then on gollum will know to automatically look there for that model grid.
+
