### 2024-05-15 3a09d43
Make download documentation more agnostic
Hedges for everyone
