+++ docs/fetching_models.rst
@@ -9,13 +9,11 @@ Downloading model grids
 Sonora
 ======
 
-Currently we only support the 2018 Sonora Model grids.  In a web browser, navigate to the `Sonora 2018 Zenodo website <https://zenodo.org/record/1309035#.YafR7oDML9A>`_.  Click on the ``Download`` button next to ``spectra.tar     1.2 GB``.  That will download the ``spectra.tar`` file to your ``~/Downloads/`` directory or equivalent default location.  Untar that directory.  The pattern should look something like below, depending on what your operating system defaults are, so copy-ing and past-ing this code is not likely to work--- try each step at a time.  ::
+Currently we only support the 2018 Sonora Model grids.  In a web browser, navigate to the `Sonora 2018 Zenodo website <https://zenodo.org/record/1309035#.YafR7oDML9A>`_.  Click on the ``Download`` button next to ``spectra.tar     1.2 GB``.  That will download the ``spectra.tar`` file to wherever you've chosen to download the spectra.  Then you want to untar ``spectra.tar``.  The pattern should look something like below, depending on what your operating system defaults are, so copying and pasting this code is not likely to work--- take it step by step. ``~/libraries/raw/Sonora`` can be replaced with wherever you want the spectral models to live long-term, but our examples here continue to use that pattern.  ::
 
     tar -xzvf spectra.tar
     mv spectra ~/libraries/raw/Sonora
     
-
-
 The end result should look something like this: ::
 
     $ ls ~/libraries/raw/Sonora/
@@ -32,19 +30,18 @@ PHOENIX
 
 .. note::
 
-    Downloading all the PHOENIX models can take hours or days! Start the downloading early.
+    Downloading all the PHOENIX models can take a ridiculous amount of time. Start the downloading as soon as you can!
 
 The PHOENIX models total over 100 GB, and generally download at a relatively slow bandwidth.  The files are arranged into sub-directories for metallicity and alpha-element abundance.
 
-You can get all of the PHOENIX models in one-fell-swoop from the command line if you have `wget` ::
+You can get all of the PHOENIX models in one-fell-swoop from the command line if you have ``wget``. Replace ``~/Downloads`` with wherever you want to download the models to. ::
 
     cd ~/Downloads
     wget -r -l 0 ftp://phoenix.astro.physik.uni-goettingen.de/HiResFITS/
 
-If you don't have `wget` on your computer, please help the ``gollum`` grow by filing a GitHub Issue with how you resolved the problem, or what problems you are encountering.
+If you don't have ``wget`` on your computer, please help the ``gollum`` grow by filing a GitHub Issue with how you resolved the problem, or what problems you are encountering.
 
-As noted, this process will take a while as each individual file is painstakingly downloaded from a single German computer.  The commandline script, as written, will preserve the directory structure--- that's good! The gollum code demands that the directory structure is preserved.  Once it's all downloaded it should look like this ::
-
+As noted, this process will take a while as each individual file is painstakingly downloaded from a single German computer.  The commandline script, as written, will preserve the directory structure--- that's good! The ``gollum code`` demands that the directory structure is preserved.  Once it's all downloaded it should look like this ::
 
     # From inside this directory: ~/Downloads/phoenix.astro.physik.uni-goettingen.de/
     $ tree 
@@ -73,8 +70,7 @@ As noted, this process will take a while as each individual file is painstakingl
         └── WAVE_PHOENIX-ACES-AGSS-COND-2011.fits
 
 
[…]
