### 2024-05-15 BrownDwarf/gollum#111 @Sujay-Shankar
I've pushed changes that make the documentation more download-agnostic. To your latter point, with PHOENIX, users are required to mess with the native directory structure to get it to how gollum likes it, so I think having users download to an intermediate location and then populate the final location with the correct structure is less prone to things overwriting themselves and breaking. If the current state is satisfactory, let me know and we can close this issue. Thanks for pointing this out!
