### 2024-05-15 2107e0d
Hedge for no interpolation
