### 2024-05-15 BrownDwarf/gollum#112 @Sujay-Shankar
That's true, we should mention up front that we don't currently support grid interpolation. It's a future plan to implement into gollum, just not one of our goals for gollum in its current iteration. I put a bit on the homepage specifying this so users don't get the wrong idea.
