### 2024-05-15 82179c4
Cut down PHOENIX docstring
- Also testing how the changes reflect in the webpage
