### 2024-05-15 58cc002
Update Sonora and Telluric docstrings
New style looks good
