### 2024-05-20 BrownDwarf/gollum#113 @Sujay-Shankar
The most recent commits should have fixed the extra headers appearing in the documentation pages, instead replaced by module-specific headers for PHOENIX, Sonora, etc. Are there any other issues that stand out in the documentation?
