### 2024-05-25 BrownDwarf/gollum#106 @DanielAndreasen
I just searched for the error and found a stackoverflow answer, but I'm unable to find it now. It's fine for me to close this issue as is, or if you want, add a small footnote about the solution I found in the docs.
