### 2024-05-26 BrownDwarf/gollum#115 @DanielAndreasen
Using dashboards
Hi again,

As you might have noticed, I'm not a big jupyter fan. I prefer to run scripts directly from the terminal, but that is not possible with the dashboard (which seems to be the coolest feature); https://gollum-astro.readthedocs.io/en/latest/tutorials/gollum_demo_Sonora_and_BDSS.html.

I was poking around, and it seems there is running some kind of service, and some calls to bokeh. Therefore it does not seem straightforward to make it work like I want. I tried in jupyter-notebook (something I haven't opened for many years, so I might have done something wrong), but couldn't get that to work either. Here are the errors I got:

```python
Processing Teff=2400K|logg=5.50|Z=0.5: 100%|███████████████████████████████████████████████████████████████████████████████████████████| 891/891 [01:03<00:00, 14.04it/s]

594 files not found; grid may not cover given parameter ranges fully

ERROR:tornado.application:Uncaught exception GET /autoload.js?bokeh-autoload-element=a25b0148-cc1f-44d2-90ff-a621dc0a2c73&bokeh-absolute-url=http://localhost:44125&resources=none (127.0.0.1)
HTTPServerRequest(protocol='http', host='localhost:44125', method='GET', uri='/autoload.js?bokeh-autoload-element=a25b0148-cc1f-44d2-90ff-a621dc0a2c73&bokeh-absolute-url=http://localhost:44125&resources=none', version='HTTP/1.1', remote_ip='127.0.0.1')
Traceback (most recent call last):
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/tornado/web.py", line 1790, in _execute
    result = await result
             ^^^^^^^^^^^^
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/server/views/autoload_js_handler.py", line 62, in get
    session = await self.get_session()
              ^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/server/views/session_handler.py", line 145, in get_session
    session = await self.application_context.create_session_if_needed(session_id, self.request, token)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/server/contexts.py", line 240, in create_session_if_needed
    self._application.initialize_document(doc)
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/application/application.py", line 190, in initialize_document
    h.modify_document(doc)
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/application/handlers/function.py", line 140, in modify_document
    self._func(doc)
  File "/home/daniel/Documents/programming/astro/gollum/src/gollum/sonora.py", line 515, in create_interact_ui
    fig = figure(
          ^^^^^^^
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/plotting/_figure.py", line 196, in __init__
    self._raise_attribute_error_with_matches(name, names | opts.properties())
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/core/has_props.py", line 379, in _raise_attribute_error_with_matches
    raise AttributeError(f"unexpected attribute {name!r} to {self.__class__.__name__}, {text} attributes are {nice_join(matches)}")
AttributeError: unexpected attribute 'plot_height' to figure, similar attributes are outer_height, height or min_height
ERROR:tornado.application:Uncaught exception GET /autoload.js?bokeh-autoload-element=a25b0148-cc1f-44d2-90ff-a621dc0a2c73&bokeh-absolute-url=http://localhost:44125&resources=none (127.0.0.1)
HTTPServerRequest(protocol='http', host='localhost:44125', method='GET', uri='/autoload.js?bokeh-autoload-element=a25b0148-cc1f-44d2-90ff-a621dc0a2c73&bokeh-absolute-url=http://localhost:44125&resources=none', version='HTTP/1.1', remote_ip='127.0.0.1')
Traceback (most recent call last):
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/tornado/web.py", line 1790, in _execute
    result = await result
             ^^^^^^^^^^^^
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/server/views/autoload_js_handler.py", line 62, in get
    session = await self.get_session()
              ^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/server/views/session_handler.py", line 145, in get_session
    session = await self.application_context.create_session_if_needed(session_id, self.request, token)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/server/contexts.py", line 240, in create_session_if_needed
    self._application.initialize_document(doc)
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/application/application.py", line 190, in initialize_document
    h.modify_document(doc)
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/application/handlers/function.py", line 140, in modify_document
    self._func(doc)
  File "/home/daniel/Documents/programming/astro/gollum/src/gollum/sonora.py", line 515, in create_interact_ui
    fig = figure(
          ^^^^^^^
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/plotting/_figure.py", line 196, in __init__
    self._raise_attribute_error_with_matches(name, names | opts.properties())
  File "/home/daniel/Documents/programming/astro/gollum/.venv/lib/python3.12/site-packages/bokeh/core/has_props.py", line 379, in _raise_attribute_error_with_matches
    raise AttributeError(f"unexpected attribute {name!r} to {self.__class__.__name__}, {text} attributes are {nice_join(matches)}")
AttributeError: unexpected attribute 'plot_height' to figure, similar attributes are outer_height, height or min_height
ERROR:tornado.access:500 GET /autoload.js?bokeh-autoload-element=a25b0148-cc1f-44d2-90ff-a621dc0a2c73&bokeh-absolute-url=http://localhost:44125&resources=none (127.0.0.1) 63.24ms
ERROR:tornado.access:500 GET /autoload.js?bokeh-autoload-element=a25b0148-cc1f-44d2-90ff-a621dc0a2c73&bokeh-absolute-url=http://localhost:44125&resources=none (127.0.0.1) 63.24ms
```

This is more than I can debug quickly. I have installed version 3.4.1 of bokeh. Any ideas what is going wrong here?

Review link: https://github.com/openjournals/joss-reviews/issues/6601
