### 2024-05-28 327de86
Add footnote for pyqt5 error
Closes #106
