### 2024-05-28 c5f8dc3
Fix bokeh figure outdated attributes in Sonora
