### 2024-05-28 BrownDwarf/gollum#115 @Sujay-Shankar
Looking at your error, it's throwing it because Sonora's Bobcat dashboard is using an outdated bokeh figure constructor that uses arguments plot_height and plot_width instead of height and width. I fixed this with the Sonora Diamondback dashboard when I tested it, but forgot to do so with Bobcat as well. This error is fixed in the recent commit, I've updated the constructor to use the correct arguments.

With regards to running the dashboard outside of jupyter notebook, I think we never anticipated use cases outside of jupyter so the dashboard as it currently is doesn't say, pop out into its own window and do stuff from there. I think you're right in that bokeh has some functionality that would let one do this, but it would require an overhaul of the way we handle dashboard objects and possibly a new script that bokeh would have to use to pop open a window that non-jupyter users could use. I'll see what I can do about this and get back to you.
