### 2024-05-28 45af05c
Fix PHOENIX demo notebook by adding data and raises-exception tag
