### 2024-05-28 34f86f8
Fix specutils versioning
