+++ .github/workflows/gollum-tests.yml
@@ -10,7 +10,7 @@ jobs:
       fail-fast: false
       matrix:
         python-version: [3.8, 3.9, 3.10.9, 3.11]
-        specutils-version: [1.8, 1.10, 1.12, 1.14]
+        specutils-version: [1.8, 1.11, 1.14]
 
     steps:
       - uses: actions/checkout@v4
