### 2024-05-28 7b208b0
Finish cleaning up documentation
Closes #113
