### 2024-05-29 BrownDwarf/gollum#115 @DanielAndreasen
I will try it again soon, and let you know if it works better now. Unfortunately I do not have much free time these days. Sorry for the late replies.
