### 2024-06-01 BrownDwarf/gollum#115 @DanielAndreasen
Okay, the dashboard works really well for me now. The only thing missing before you can close this, is to update the notebook. Right now it has an `AssertionError` in the documentation, which does not happen. So I guess you just have to run the notebook again, and commit it.
