### 2024-06-01 BrownDwarf/gollum#117 @DanielAndreasen
PHOENIX dashboard
This is related to #116.

* I think it needs to say where to get the example data. I found it here: https://github.com/OttoStruve/muler_example_data/tree/main/IGRINS/01_IGRINS_test_data
And mention all the three files that are needed for this
* Currently I get an error when creating the grid, since no points are found. I will try to debug, to make sure I didn't do anything wrong.

Review link: https://github.com/openjournals/joss-reviews/issues/6601
