### 2024-06-01 BrownDwarf/gollum#117 @DanielAndreasen
For the grid, the structure of the PHOENIX models were slightly wrong (I will have to check if it is written correctly in the documentation a bit later, or if I did something wrong).
However, now the dashboard crashes due to some problems with the wavelength range. If I change the grid to: `grid = PHOENIXGrid(teff_range=(2500, 7000), logg_range=(2, 5), Z_range=(0, 0.5), wl_lo=21350, wl_hi=21590)` it works (I changed `wl_lo/wl_hi`).
This is not as smooth as the other dashboard, but I guess that is due to the longer wavelength range.

I think I nice addition to both examples is to write the actual parameters, Teff, logg, etc. so the user can "verify" their best fit. At least this is something I would like to have.
