### 2024-06-04 BrownDwarf/gollum#117 @Sujay-Shankar
I'll take a look at all these issues and let you know when they've been fixed.
And yes, the dashboard does slow down as the wavelength range increases. A trick I've found to make it more responsive in such cases is to click through the sliders instead of dragging. This prevents the callbacks from stacking up and slowing the dashboard down.
