### 2024-06-05 5bad3e3
Update Sonora demo and data download docs
