+++ docs/fetching_models.rst
@@ -23,6 +23,8 @@ The end result should look something like this: ::
     sp_t2100g31nc_m0.0.gz    sp_t450g562nc_m0.0.gz    sp_t950g56nc_m0.0.gz
                 [...]
 
+Due to the way the Sonora models are packaged, models with solar metallicity end in the ``.gz`` file extension, while other models have no file extension. Confirm that this is the case in your system. We know that this is somewhat inconsistent and are looking into ways to repackage model grids for express ``gollum`` usage in the future.
+
 
 PHOENIX
 =======
@@ -41,7 +43,7 @@ You can get all of the PHOENIX models in one-fell-swoop from the command line if
 
 If you don't have ``wget`` on your computer, please help ``gollum`` grow by filing a GitHub Issue with how you resolved the problem, or what problems you are encountering.
 
-As noted, this process will take a while as each individual file is painstakingly downloaded from a single German computer.  The commandline script, as written, will preserve the directory structure--- that's good! The ``gollum code`` demands that the directory structure is preserved.  Once it's all downloaded it should look like this ::
+As noted, this process will take a while as each individual file is painstakingly downloaded from a single German computer. The commandline script, as written, will preserve the directory structure--- that's good! The ``gollum code`` demands that the directory structure is preserved. Once it's all downloaded it should look like this ::
 
     # From inside this directory: ~/Downloads/phoenix.astro.physik.uni-goettingen.de/
     $ tree 
@@ -69,22 +71,10 @@ As noted, this process will take a while as each individual file is painstakingl
         │    └── Z-4.0
         └── WAVE_PHOENIX-ACES-AGSS-COND-2011.fits
 
-
-While you `could` leave these directories in say ``~/Downloads/phoenix.astro.physik.uni-goettingen.de/``, we recommend making a more permanent and recognizable home for these models.  In particular ``gollum`` attempts to search a single default path for models: ``~/libraries/raw/``, which we'll run with for the sake of the examples, but you can also store the models somewhere else and edit ``gollum``'s configs to default to your specified location (see below). ::
-
-    mv ~/Downloads/phoenix.astro.physik.uni-goettingen.de/HiResFITS/PHOENIX-ACES-AGSS-COND-2011 ~/libraries/raw/PHOENIX/
-    ls ~/libraries/raw/PHOENIX/
-    Z-0.0  Z+0.5  Z-0.5  Z+1.0  Z-1.0  Z-1.5  Z-2.0  Z-3.0  Z-4.0
-
-Finally, you must copy the wavelength file into this directory as well.  Notice that this placement breaks the native directory structure, so you must complete this step in order for ``gollum`` to work. ::
-
-    mv ~/Downloads/phoenix.astro.physik.uni-goettingen.de/HiResFITSWAVE_PHOENIX-ACES-AGSS-COND-2011.fits ~/libraries/raw/PHOENIX/
-    ls ~/libraries/raw/PHOENIX/
-    WAVE_PHOENIX-ACES-AGSS-COND-2011.fits  Z-0.0  Z+0.5  Z-0.5  Z+1.0  Z-1.0  Z-1.5  Z-2.0  Z-3.0  Z-4.0
-
-You can optionally set user defaults for paths for each model grid. First, go into ``gollum``'s source directory and look for a file called ``config.env`` (NOT ``config_template.env``, do not edit this file!). If it doesn't exist, then run the following: ::
+You can set user defaults for paths for each model grid. First, go into ``gollum``'s source directory and look for a file called ``config.env`` (NOT ``config_template.env``, do not edit this file!). If it doesn't exist, then run the following: ::
 
     cp config_template.env config.env
 
[…]
