### 2024-06-05 BrownDwarf/gollum#115 @Sujay-Shankar
The notebook has been rerun and committed, just some tiny commits left until the errors are cleaned up in full. To your earlier point, I think that as gollum currently is and has been, the dashboard was something that required jupyter. While I do agree with you and want to make jupyter-less functionality possible, I'd like to put that up in a future version release where we can do the major design overhauls and keep everything clean instead of trying to hack it into the current release. Thoughts?
