### 2024-06-05 BrownDwarf/gollum#115 @Sujay-Shankar
I've created a standalone issue about it so that it's visible outside of the JOSS review process, so we can safely close this one.
