### 2024-06-05 BrownDwarf/gollum#117 @Sujay-Shankar
Alright, I've changed the docs a little bit, they now give links to both the spec_a0v.fits and variance.fits files that IGRINSSpectrum needs to initialize (although gollum's repo now comes with them preloaded in the tutorial_data directory, which the docs also reflect now). The structure of the PHOENIX grid was a little inconsistent and confusing in the docs, I've updated them to be much easier to get up and running with little to no modification required after the download is complete. Basically all your config.env needs to point to is the directory which contains the wavelength file and the PHOENIX-ACES-AGSS-COND-2011 file (this specifically housing all the metallicity slices).

I've also added the verification for the best fit parameters (it is by eye, so not spot on obviously, but probably in the ballpark).
