### 2024-06-28 BrownDwarf/gollum#119 @DanielAndreasen
Guidelines for contribution and statement of need
I'm just about done with the checklist before I can recommend this for publication.

Two things missing from that checklist are:
* Guidelines for contribution. Simply add in the readme how to contribute, either to the code (pull request), or with questions (open an issue).
* Statement of need. This is more or less written in the paper, but can you add a section with that title? I.e. what is the need of this tool (already written; for education and explorative research basically).

https://github.com/openjournals/joss-reviews/issues/6601
