### 2024-06-28 #6601 @DanielAndreasen
> Hi @rhdtownsend & @DanielAndreasen, I just wanted to ask how the review is progressing. It looks like @Sujay-Shankar has responded to all the issues raised in the Gollum repo and I didn't see any more left open. Could you continue down your list and raise any further issues?

Yup, almost done. Just opened the last minor issue. After that is solved, I will recommend this tool for publication.
