### 2024-06-28 84e0937
Tweaks to readme and statement of need in paper
