+++ paper/paper.md
@@ -44,7 +44,7 @@ The spectra of stars, brown dwarfs, and planets are complex and information-rich
 
 # Statement of need
 
-`gollum` is a Python package for intuitive analysis and visualization of precomputed synthetic spectra. Its API is designed to have modules dedicated to each model grid it supports, with each module then containing classes for both individual spectra and bulk grid access. The programmatic interface to spectral analysis uses method-chaining to make `gollum` code very readable, taking inspiration from frameworks like lightkurve [@lightkurve:2018]. The visual interface in the form of interactive dashboards powered by `bokeh` [@bokeh2018] sports low-latency sliders and toggles. This allows users to tweak both fundamental stellar parameters and extrinsic parameters such as radial velocity and rotational broadening, building their intuition. `gollum`'s modularity allows for a wide range of model grids to potentially be supported, and its performance is optimized with libraries such as `numpy`, `scipy`, `astropy`, and `specutils` to allow for quick loading and processing of large amounts of data [@harris:2020; @virtanen:2020; @astropy:2022; @earl:2023].
+`gollum` is a Python package for intuitive analysis and visualization of precomputed synthetic spectra, targeted towards usage for education and exploratory research. Its API is designed to have modules dedicated to each model grid it supports, with each module then containing classes for both individual spectra and bulk grid access. The programmatic interface to spectral analysis uses method-chaining to make `gollum` code very readable, taking inspiration from frameworks like lightkurve [@lightkurve:2018]. The visual interface in the form of interactive dashboards powered by `bokeh` [@bokeh2018] sports low-latency sliders and toggles. This allows users to tweak both fundamental stellar parameters and extrinsic parameters such as radial velocity and rotational broadening, building their intuition. `gollum`'s modularity allows for a wide range of model grids to potentially be supported, and its performance is optimized with libraries such as `numpy`, `scipy`, `astropy`, and `specutils` to allow for quick loading and processing of large amounts of data [@harris:2020; @virtanen:2020; @astropy:2022; @earl:2023].
 
 `gollum` appeals to use cases ranging from entry-level astronomers to seasoned researchers with its combination of intuition-building and performance. The framework has been demonstrated on high-resolution spectra of brown dwarfs from Keck-NIRSPEC using `bdexda` [@kimani-stewart:2021], has experimental support for starspots, used in `acdc` [@cao:2022], and its programmatic interface has been used in the analysis of IGRINS spectra with `plotspec` [@kaplan2023]. It interoperates with `muler` [@gully-santiago:2022a], a similar framework designed for observed data from échelle spectrographs, and can also be used for flux calibration and empirical telluric correction. `gollum`'s programmatic interface is being used to create an extension to the `blase` framework [@gully-santiago:2022b; @shankar:2023] that will allow for inference of fundamental stellar parameters from observed spectra using interpretable machine learning and interpolation techniques, taking inspiration from other frameworks such as `starfish` [@czekala:2015] that also specialize in spectroscopic inference. `gollum` currently supports PHOENIX, Sonora Bobcat, and Sonora Diamondback, model grids, and implementations for other model grids such as CoolTLUSTY and Sonora Elf Owl are in progress [@husser:2013; @marley:2021; @morley:2024; @lacy:2023; @mukherjee:2024].
 
