### 2024-06-28 BrownDwarf/gollum#119 @Sujay-Shankar
Thanks for the feedback Daniel!
We already had the contribution instructions in the readme, and the Statement of Need section was in the paper, but I suspect you meant that they should be more explicit.
I've updated both to incorporate both of your points, the readme is more exhaustive in its contribution instructions and the statement of need section now explicitly states that gollum is targeted towards education and explorative research.
Let me know if the changes in the recent commit are satisfactory, and we should be good to go ahead and close this one up.
