### 2024-06-28 BrownDwarf/gollum#119 @DanielAndreasen
Beautiful. I think I might just have overlooked some things during my first review of the readme/paper. But it reads better now I think.
