### 2024-07-26 #6601 @Sujay-Shankar
Hi all, just wanted to check in with @rhdtownsend and see if there were any more updates/issues to raise regarding gollum.
