### 2024-07-29 #6601 @rhdtownsend
Sorry for the delay in responding — been on a fair bit of travel recently. I’ve no more suggestions, and happy to recommend publication proceeds.

> On Jul 26, 2024, at 2:27 AM, Sujay Shankar ***@***.***> wrote:
>
>
> Hi all, just wanted to check in with @rhdtownsend and see if there were any more updates/issues to raise regarding gollum.
