### 2024-07-29 #6601 @warrickball
Thanks @DanielAndreasen and @rhdtownsend for your reviews.

@Sujay-Shankar, the final things to do are now specified under the **Additional Author Tasks After Review is Complete** above. (I'll work through the editor tasks.) The main thing is to make a new tagged release of the software, upload the archive to a repository like Zenodo and check that the metadata is correct. I'll then record the DOI and release version here ahead of the paper being published in JOSS.
