### 2024-07-30 3b4b674
Reflect 0.4.0 in packaging files
