### 2024-07-30 #6601 @Sujay-Shankar
Alright, gollum's latest version is now 0.4.1, published as a release on both PyPI and GitHub. I've turned on Zenodo's GitHub integration, so 0.4.1 has a Zenodo archive under the following DOI: 10.5281/zenodo.13132601. The author list matches and both are using the same MIT license. Let me know if that all looks good!
