### 2024-07-30 #6601 @warrickball
The archive title appears to be

> **BrownDwarf/gollum: v0.4.1: Kickstarting Zenodo sync**

which probably isn't what you want. You can probably amend the title directly in the Zenodo archive to remove (or replace) the "Kickstarting Zenodo sync" part.
