### 2024-07-31 #6601 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
