### 2024-08-04 #6601 @dfm
@Sujay-Shankar — I've opened a PR with some small edits to the bibliography. Please take a look at that and I'll do the final processing once that's in!
