### 2024-08-04 #6601 @Sujay-Shankar
Merge approved! Good catch, I forgot about the ApJ latex thing.
