### 2024-08-09 #6601 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
