### 2024-08-09 #6601 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
